// sdk/nodejs/ec2/enums.ts
export const Tenancy = {
    Default: "default",
    Dedicated: "dedicated",
} as const;

export type Tenancy = (typeof Tenancy)[keyof typeof Tenancy];